requirements2.txt
# python-social-auth
MySQL-python
pinax-theme-bootstrap==5.4.0
django-user-accounts==1.0c9
metron==1.3.1
eventlog==0.8.0
django-jsonfield==0.9.12
python-social-auth==0.2.1
singledispatch
unicodecsv